+++ packages/search-form/dts/search-form/utils/index.d.ts
@@ -1,0 +1,4 @@
+import { IField, IFieldEvent } from "../types/index";
+export declare const isEmpty: (value: any) => boolean;
+export declare const initDefaultRow: (field: IFieldEvent) => void;
+export declare const initDefaultData: (fields: IField) => IField;