+++ packages/search-form/src/search-form/types/index.ts
@@ -18,8 +18,7 @@ export type IGetOptionforKey = (optionsValue: string | number) => IOptions;
 
 export type IGetOptionLabelForKey = (optionsValue: string | number) => string;
 
-export type IFetchDict = (optionsValue?: string) => Promise<IOptions[]>;
-
+export type IFetchDict = Promise<IOptions[]>
 
 export type IFieldEventValue = number | string | string[] | number[];