+++ example/CHANGELOG.md
@@ -1,5 +1,68 @@
 # example
 
+## 1.2.18
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.18
+
+## 1.2.17
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.17
+
+## 1.2.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.16
+
+## 1.2.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.15
+
+## 1.2.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.14
+
+## 1.2.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.13
+
+## 1.2.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.12
+
+## 1.2.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.11
+
+## 1.2.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @tempoi/search-form@1.2.10
+
 ## 1.2.9
 
 ### Patch Changes